requirements.txt
anyio==4.7.0
APScheduler==3.11.0
asyncio==3.4.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==21.9
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tzlocal==5.2
urllib3==2.3.0
